+++ src/client/java/techreborn/client/gui/GuiAESU.java
@@ -85,7 +85,7 @@ public class GuiAESU extends GuiBase<BuiltScreenHandler> {
 			matrices.pop();
 		}
 
-		builder.drawMultiEnergyBar(drawContext, this, 81, 28, (int) blockEntity.getEnergy(), (int) blockEntity.getMaxStoredPower(), mouseX, mouseY, 0, layer);
+	builder.drawMultiEnergyBar(drawContext, this, 81, 28, blockEntity.getEnergy(), blockEntity.getMaxStoredPower(), mouseX, mouseY, 0, layer);
 	}
 
 	public void onClick(int amount) {
