+++ src/client/java/techreborn/client/gui/GuiAutoCrafting.java
@@ -55,7 +55,7 @@ public class GuiAutoCrafting extends GuiBase<BuiltScreenHandler> {
 		final Layer layer = Layer.FOREGROUND;
 
 		builder.drawProgressBar(drawContext, this, blockEntityAutoCraftingTable.getProgress(), blockEntityAutoCraftingTable.getMaxProgress(), 120, 44, mouseX, mouseY, GuiBuilder.ProgressDirection.RIGHT, layer);
-		builder.drawMultiEnergyBar(drawContext, this, 9, 26, (int) blockEntityAutoCraftingTable.getEnergy(), (int) blockEntityAutoCraftingTable.getMaxStoredPower(), mouseX, mouseY, 0, layer);
+	builder.drawMultiEnergyBar(drawContext, this, 9, 26, blockEntityAutoCraftingTable.getEnergy(), blockEntityAutoCraftingTable.getMaxStoredPower(), mouseX, mouseY, 0, layer);
 	}
 
 	@Override
