+++ src/client/java/techreborn/client/gui/GuiBatbox.java
@@ -68,6 +68,6 @@ public class GuiBatbox extends GuiBase<BuiltScreenHandler> {
 			matrices.pop();
 		}
 
-		builder.drawMultiEnergyBar(drawContext, this, 81, 28, (int) blockEntity.getEnergy(), (int) blockEntity.getMaxStoredPower(), mouseX, mouseY, 0, layer);
+	builder.drawMultiEnergyBar(drawContext, this, 81, 28, blockEntity.getEnergy(), blockEntity.getMaxStoredPower(), mouseX, mouseY, 0, layer);
 	}
 }
