+++ src/client/java/techreborn/client/gui/GuiBlastFurnace.java
@@ -82,7 +82,7 @@ public class GuiBlastFurnace extends GuiBase<BuiltScreenHandler> {
 			addHologramButton(76, 56, 212, layer).clickHandler(this::onClick);
 			builder.drawHologramButton(drawContext, this, 76, 56, mouseX, mouseY, layer);
 		}
-		builder.drawMultiEnergyBar(drawContext, this, 9, 19, (int) blockEntity.getEnergy(), (int) blockEntity.getMaxStoredPower(), mouseX, mouseY, 0, layer);
+	builder.drawMultiEnergyBar(drawContext, this, 9, 19, blockEntity.getEnergy(), blockEntity.getMaxStoredPower(), mouseX, mouseY, 0, layer);
 	}
 
 	public void onClick(GuiButtonExtended button, Double x, Double y) {
