+++ src/main/java/techreborn/items/tool/ChainsawItem.java
@@ -41,14 +41,14 @@ import reborncore.common.util.ItemUtils;
 
 public class ChainsawItem extends AxeItem implements RcEnergyItem {
 
-	public final int maxCharge;
-	public final int cost;
+	public final long maxCharge;
+	public final long cost;
 	public final float poweredSpeed;
 	private final float unpoweredSpeed;
 	public final Item referenceTool;
 	public final RcEnergyTier tier;
 
-	public ChainsawItem(ToolMaterial material, int energyCapacity, RcEnergyTier tier, int cost, float poweredSpeed, float unpoweredSpeed, Item referenceTool) {
+	public ChainsawItem(ToolMaterial material, long energyCapacity, RcEnergyTier tier, long cost, float poweredSpeed, float unpoweredSpeed, Item referenceTool) {
 		// combat stats same as for diamond axe. Fix for #2468
 		super(material, 5.0F, -3.0F, new Item.Settings().maxCount(1).maxDamage(-1));
 		this.maxCharge = energyCapacity;
@@ -59,7 +59,7 @@ public class ChainsawItem extends AxeItem implements RcEnergyItem {
 		this.referenceTool = referenceTool;
 	}
 
-	public int getCost() {
+	public long getCost() {
 		return cost;
 	}
